Consolidate socklen_t typedef to fut_socket.h
Move socklen_t typedef from scattered syscall files to the central
fut_socket.h header. This eliminates duplicate definitions and ensures
consistent type usage across all socket-related syscalls.

Also fixes sign comparison warnings in FD upper bounds checks by
removing unnecessary unsigned int casts when comparing against max_fds.

Files cleaned up:
- sys_accept.c, sys_bind.c, sys_connect.c, sys_listen.c
- sys_getsockopt.c, sys_setsockopt.c
- sys_getsockname.c, sys_getpeername.c
- sys_sendto.c, sys_recvfrom.c

+++ kernel/sys_listen.c
@@ -81,8 +81,8 @@ long sys_listen(int sockfd, int backlog) {
     }
 
     /* Phase 5: Validate fd upper bounds to prevent out-of-bounds access */
-    if ((unsigned int)local_sockfd >= task->max_fds) {
-        listen_printf("[LISTEN] listen(sockfd=%d, backlog=%d) -> EBADF (fd exceeds max_fds %u)\n",
+    if (local_sockfd >= task->max_fds) {
+        listen_printf("[LISTEN] listen(sockfd=%d, backlog=%d) -> EBADF (fd exceeds max_fds %d)\n",
                    local_sockfd, local_backlog, task->max_fds);
         return -EBADF;
     }
